Fix AVRS patch rpm download, Fix AVRS os repo dep
Yum was not keeping the downloaded rpms in /6wind repo,
it was removed after the next yum transaction.
"The downloaded packages were saved in cache until the next successful
transaction.". Using yumdownloader seems to fix it.

Avrs depends on python3-pyroute2, this is a package in
openstack 16.1 rpms. So this repository needs to be enabled.

Change-Id: I8ca6aef3bbe3d356289644ff553b43360758d9c1
Closed-Bug: #OPENSTACK-2970
Closed-Bug: #OPENSTACK-2968