Update doc to use hostname for AVRS container pull script
With default rhosp installations the overcloud will be
configured to use the undercloud regstriy using the hostname.
In the nuage AVRS container image pull script it is documented that it
is fine to use the IP. This is wrong. It should use the undercloud
registry using hostname as well, otherwise the overcloud deploy might
not be able to download them.

Change-Id: I43575b877071899de214eedfdcc2beb0ce64c134